+++ packages/ethers-provider-fireblocks/src/constants.ts
@@ -1,3 +1,9 @@
 import { DEBUG_NAMESPACE } from '@fireblocks/fireblocks-web3-provider'
 
 export const DEBUG_NAMESPACE_SILENTDATA_INTERCEPTOR = `${DEBUG_NAMESPACE}:silentdata-interceptor`
+
+export const EIP712Domain = [
+  { name: 'name', type: 'string' },
+  { name: 'version', type: 'string' },
+  { name: 'chainId', type: 'uint256' },
+]