+++ packages/ethers-provider/src/constants.ts
@@ -1,9 +1,1 @@
-export const SIGN_RPC_METHODS = [
-  'eth_getTransactionByHash',
-  'eth_getBalance',
-  'eth_getTransactionCount',
-  'eth_getProof',
-  'eth_getTransactionReceipt',
-]
-
 export const DEBUG_NAMESPACE = 'silentdata:ethers-provider'