+++ packages/ethers-provider-fireblocks/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @appliedblockchain/silentdatarollup-ethers-provider-fireblocks
 
+## 1.0.4
+
+### Patch Changes
+
+- Chain id added as domain separator
+  Packages improvements
+  Examples improvements and fixes
+- Updated dependencies
+  - @appliedblockchain/silentdatarollup-core@1.0.4
+
 ## 1.0.3
 
 ### Patch Changes